mrd-viz: split webview viewer.ts into focused media modules
Addresses Yulia's PR #81 review: the webview refactor left most logic in one ~1024-line media/viewer.ts. Split it into a slim entry point plus focused modules under media/viewer/ (state, dom, slice, messaging, viewport, metadata, selectedTile, mosaic). esbuild --bundle still emits a single media/viewer.js; behavior is unchanged.
