+++ mrd-viz/extension/mrd-viz/media/tsconfig.json
@@ -18,6 +18,7 @@
 		"skipLibCheck": true
 	},
 	"include": [
-		"viewer.ts"
+		"viewer.ts",
+		"viewer/**/*.ts"
 	]
 }
